+++ Cargo.toml
@@ -8,12 +8,14 @@ license = "MIT"
 repository = "https://github.com/ribru17/ts_query_ls"
 
 [dependencies]
+clap = { version = "4.5.31", features = ["derive"] }
 dashmap = "6.1.0"
 dissimilar = "1.0.9"
 env_logger = "0.11.5"
 lazy_static = "1.5.0"
 libloading = "0.8.5"
 log = "0.4.22"
+rayon = "1.10.0"
 regex = "1.11.0"
 ropey = "1.6.1"
 serde = "1.0.210"
@@ -24,6 +26,7 @@ tower-lsp = "0.20.0"
 tracing-subscriber = "0.3.18"
 tree-sitter = { version = "0.25.0", features = ["std", "wasm"] }
 tree-sitter-query = { git = "https://github.com/tree-sitter-grammars/tree-sitter-query", rev = "9bf4d5cad5ecf51d9102cd50765e8d6119616a9a", version = "0.4.0" }
+walkdir = "2.5.0"
 
 [build-dependencies]
 cc = "1.1.30"
